requirements.txt
Pillow==2.6.1
RPi.GPIO==0.6.3
argparse==1.2.1
arrow==0.10.0
chardet==2.3.0
colorama==0.3.2
html5lib==0.999
inkyphat==0.0.2
ndg-httpsclient==0.3.2
numpy==1.8.2
pyOpenSSL==0.13.1
pyasn1==0.1.7
python-apt==0.9.3.12
python-dateutil==2.6.1
pytz==2017.2
requests==2.4.3
scrollphathd==0.0.1
six==1.8.0
smbus==1.1
spidev==3.2
urllib3==1.9.1
wheel==0.24.0
wsgiref==0.1.2
